New ItemList creation now occurs on OverviewScreen. This was changed due to the headache that was being cause from previous method when creating a new list an trying to add items right away. AlertDialog would appear again when navigating back from AddScreen.
